% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esr_dissolved}
\alias{esr_dissolved}
\title{List of Dissolved French University}
\format{A data frame with 2 variables:
 \describe{
  \item{id}{A \code{\link{character}} string giving the wikidata id.}
  \item{alias}{A \code{\link{character}} string giving the name of the
  university.}
 }}
\usage{
esr_dissolved
}
\description{
A dataset containing French universities that disappeared in the 1970s.
}
\seealso{
Other datasets: \code{\link{esr_level}},
  \code{\link{esr_status}}
}
\concept{datasets}
\keyword{datasets}
