% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esr_level}
\alias{esr_level}
\title{Levels of French ESR Institution Status}
\format{A data frame with 4 variables:
 \describe{
  \item{level}{an integer to id the level.}
  \item{label}{the label of the level.}
  \item{description}{a description of the level.}
  \item{example}{some example of status of this level.}
 }}
\usage{
esr_level
}
\description{
A dataset containing arbitrary levels to institutions statuts.
Used to set sizes when plotting things.
}
\seealso{
Other datasets: \code{\link{esr_dissolved}},
  \code{\link{esr_status}}
}
\concept{datasets}
\keyword{datasets}
