% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esr_merged}
\alias{esr_merged}
\title{List of Merged French Universities}
\format{A data frame with 2 variables:
 \describe{
  \item{id}{A \code{\link{character}} string giving the wikidata id.}
  \item{alias}{A \code{\link{character}} string giving the name of the
  university.}
 }}
\usage{
esr_merged
}
\description{
A dataset containing French universities that were merged in the 2000s-2010s.
}
\seealso{
Other datasets: \code{\link{esr_dissolved}},
  \code{\link{esr_level}}, \code{\link{esr_status}}
}
\concept{datasets}
\keyword{datasets}
