% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esr_status}
\alias{esr_status}
\title{Status of French Higher Education Institutions}
\format{A data frame with 6 variables:
 \describe{
   \item{id}{Wikipedia ID of the item.}
   \item{label}{the label of the item.}
   \item{deprecated}{whether this item is of recommanded use.}
   \item{level}{level of the item (see \code{\link{esr_level}}).}
   \item{wikipedia}{URL to the wikipedia notice.}
   \item{note}{note to help the user.}
 }}
\source{
\url{https://www.wikidata.org}
}
\usage{
esr_status
}
\description{
A dataset containing the main status of French higher education
institutions, together with some additionnal informations.
}
\seealso{
Other datasets: \code{\link{esr_level}}
}
\concept{datasets}
\keyword{datasets}
