% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openESR-package.R
\docType{package}
\name{openESR-package}
\alias{openESR-package}
\alias{_PACKAGE}
\alias{openESR}
\title{wikidataESR: From wikidata to R, for french ESR}
\description{
This package uses data about the french
    university system (ESR) to build and plot graphs. Wikidata is a
    convenient tool to model universities information. This package
    intends to help using data about the french universities, namely
    Enseignement Supérieur et Recherche (ESR). Its main purpose is to plot
    graphs about ESR universities relationship.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab openESR \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 0.1.0 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/xxx}{xxx}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/nfrerebeau/DataESR}
  \item Report bugs at \url{http://github.com/nfrerebeau/DataESR/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
 Julien Gossa \tab \emph{Université de Strasbourg, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
