% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.esr_graph}
\alias{plot.esr_graph}
\title{Plot}
\usage{
\method{plot}{esr_graph}(x, layout = "kamadakawai", edge_label = TRUE,
  node_label = "alias", node_dates = FALSE, node_type = "text",
  size_guide = FALSE)
}
\arguments{
\item{layout}{The layout to use to plot the graph as in
\code{\link[sna]{gplot.layout}}.}

\item{edge_label}{TRUE to plot dates on edges (default to "TRUE").}

\item{node_label}{Define the label for the nodess. Either "alias",
"alias_date", "long", or "long_date" (default to "alias").}

\item{node_type}{Define the type of drawing for the nodes. Either "text",
"text_repel", "label", or "label_repel" (default to "text").}

\item{size_guide}{TRUE to plot the guide for sizes (defalut to "FALSE").}

\item{node_date}{TODO.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Renders a graph.
}
\author{
J. Gossa, N. Frerebeau
}
