% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{wdesr_get_data}
\alias{wdesr_get_data}
\title{Load Data of a Set of Universities}
\usage{
wdesr_get_data(id, simplify = TRUE, verbose = getOption("verbose"))
}
\arguments{
\item{id}{A \code{\link{character}} vector specifying the wikidata IDs
to be accessed.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a data frame (default to \code{TRUE})?}

\item{verbose}{A \code{\link{logical}} scalar: should diagnostic messages be
generated?}
}
\value{
If \code{simplify} is \code{TRUE}, returns a
 \code{\link{data.frame}}, else returns a \code{\link{list}}.
}
\description{
Load Data of a Set of Universities
}
\examples{
\donttest{
## Warning: an active internet connection is needed
## Get the data regarding Strasbourg University
unistra <- wdesr_get_data("Q61716176")

## Get data for Clermont and Bordeaux universities
clermont <- wdesr_get_data(c("Q3551576", "Q20791505"))
}
}
\references{
\href{https://www.wikidata.org}{Wikidata}.
}
\author{
J. Gossa, N. Frerebeau
}
