% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{wdesr_get_item}
\alias{wdesr_get_item}
\title{Load Data of a University}
\usage{
wdesr_get_item(id, verbose = getOption("verbose"))
}
\arguments{
\item{id}{A \code{\link{character}} string specifying the the wikidata id of
the university.}

\item{verbose}{A \code{\link{logical}} scalar: should a diagnostic message be
generated?}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Load Data of a University
}
\examples{
\donttest{
## Warning: an active internet connection is needed
## Get the data regarding Strasbourg University
unistra <- wdesr_get_data("Q61716176")

## Get data for Clermont and Bordeaux universities
clermont <- wdesr_get_data(c("Q3551576", "Q20791505"))
}
}
\references{
\href{https://www.wikidata.org}{Wikidata}.
}
\seealso{
\link{wdesr_get_data}
}
\author{
J. Gossa, N. Frerebeau
}
\keyword{internal}
