% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/utilities.R
\name{wikidata}
\alias{wikidata}
\alias{get_item_label}
\alias{get_item_alias}
\alias{get_statement_year}
\alias{get_statement_list}
\alias{get_statement_qualifier}
\alias{get_item_label.wikidata}
\alias{get_item_alias.wikidata}
\alias{get_statement_year.wikidata}
\alias{get_statement_list.wikidata}
\alias{get_statement_qualifier.wikidata}
\title{Getters for Wikidata}
\usage{
get_item_label(item, ...)

get_item_alias(item, ...)

get_statement_year(item, ...)

get_statement_list(item, ...)

get_statement_qualifier(item, ...)

\method{get_item_label}{wikidata}(item)

\method{get_item_alias}{wikidata}(item)

\method{get_statement_year}{wikidata}(item, property)

\method{get_statement_list}{wikidata}(item, property)

\method{get_statement_qualifier}{wikidata}(item, property, qualifier)
}
\arguments{
\item{item}{An object of class \code{\link[wikidata]{WikidataR}}.}

\item{property}{A \code{\link{character}} string giving the property to get
the data from.}

\item{qualifier}{A \code{\link{character}} string giving the qualifier to
be retrieved.}
}
\value{
\code{get_item_label} returns a \code{\link{character}} string
 (look for french label first, and return english otherwise).

 \code{get_item_alias} returns a \code{\link{character}} string
 (look for shortest french label/alias).

 \code{get_statement_year} returns a \code{\link{character}} string giving
 the year stated in the item.

 \code{get_statement_list} returns a \code{\link{list}} of statements.

 \code{get_statement_list} returns a \code{\link{list}} of item statement
 qualifiers.
}
\description{
\code{get_item_label} gets the label of a \code{wikidata} item.
}
\details{
\code{get_item_alias} gets the alias of a \code{wikidata} item.

\code{get_statement_year} gets the year from a statement of a
\code{wikidata} item.

\code{get_statement_list} gets the list of statements for a given
property.

\code{get_statement_qualifier} gets the qualifiers of statements from a
\code{wikidata} item.
}
\examples{
\donttest{
## Warning: an active internet connection is needed
item <- WikidataR::get_item("Q13342")

## Name
get_item_label(item)
## Alias
get_item_alias(item)
## Foundation date
get_statement_year(item, "P571")
## Formerly known as
get_statement_list(item, "P1365")
## Predecessor of
get_statement_qualifier(item, "P1365", "P585")
}
}
\seealso{
\code{\link[WikidataR]{WikidataR}}
}
\author{
J. Gossa, N. Frerebeau
}
\keyword{internal}
