% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikidataESR-package.R
\docType{package}
\name{WikidataESR-package}
\alias{WikidataESR-package}
\alias{_PACKAGE}
\alias{WikidataESR}
\title{WikidataESR: From Wikidata to R, for French Higher Education System}
\description{
This package uses data about the french
    university system (ESR) to build and plot graphs. Wikidata is a
    convenient tool to model universities information. This package
    intends to help using data about the french universities, namely
    Enseignement Supérieur et Recherche (ESR). Its main purpose is to plot
    graphs about ESR universities relationship.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab WikidataESR \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 0.1.0 \cr
 \strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/nfrerebeau/DataESR}
  \item Report bugs at \url{http://github.com/nfrerebeau/DataESR/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
 Julien Gossa \tab \emph{Université de Strasbourg, France} \cr
}

\strong{Package maintainer}

Julien Gossa\cr
\email{gossa@unistra.fr}

ICube (UMR 7357)\cr
300, boulevard Sébastien Brant\cr
CS 10413\cr
F-67412 Illkirch cedex\cr
France
}
\keyword{internal}
