% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/wikidata.R
\name{wikiesr}
\alias{wikiesr}
\alias{get_item_status}
\alias{get_item_status.wikidata}
\title{Get the Status of a Wikidata Item}
\usage{
get_item_status(item, ...)

\method{get_item_status}{wikidata}(item)
}
\arguments{
\item{item}{An object of class \code{\link[WikidataR]{WikidataR}}.}
}
\value{
A \code{\link{data.frame}}.
}
\description{
\code{get_item_status} gets the *legal status* of a French higher education
institution.
}
\note{
It is based on the "instance of" property of Wikidata (P31) and
converted thanks to an internal dataset (see \code{\link{esr_status}}).
}
\examples{
\donttest{
## Warning: an active internet connection is needed
item <- WikidataR::get_item("Q13342")

get_item_status(item)
}
}
\seealso{
\link{esr_status}
}
\author{
J. Gossa, N. Frerebeau
}
\keyword{internal}
