% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{wdesr_get_graph}
\alias{wdesr_get_graph}
\title{Get a graph of universities.}
\usage{
wdesr_get_graph(id, property, depth = 3, active_only = FALSE,
  stop_at = c("EPST"))
}
\arguments{
\item{id}{The wikidata id of the root.}

\item{property}{The set of properties to follow.}

\item{depth}{The depth of the graph (more or less) (default to 3).}

\item{active_only}{TRUE to filter dissolved universities (default to FALSE).}

\item{stop_at}{A list of type of nodes that must not be visited furthermore
(default to "EPST").}
}
\value{
A \code{\link{list}} of edges and vertices.
}
\description{
From a root wikipedia id, the function follows a given set of properties,
building vertice and edges along the way.
}
\details{
Data are cached: use \code{\link{wdesr_clear_cache}} to refresh data from wikidata.
}
\examples{
g <- wdesr_get_graph("Q61716176", c('composante','associé'), 1)
g$edges
g$vertice
}
\author{
J. Gossa, N. Frerebeau
}
