% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/utilities.R
\name{wikiesr}
\alias{wikiesr}
\alias{get_item_status}
\alias{get_item_status.wikidata}
\title{Get the ESR Status of a wikidata Item}
\usage{
get_item_status(item, ...)

\method{get_item_status}{wikidata}(item)
}
\arguments{
\item{item}{An object of class \code{\link[wikidata]{WikidataR}}.}
}
\value{
The status the item.
}
\description{
\code{get_item_status} gets the *legal status* of an ESR institution.
}
\note{
It is based on the instance_of property (P31) and converted thanks to a
local dataset.
}
\examples{
\donttest{
## Warning: an active internet connection is needed
item <- WikidataR::get_item("Q13342")

get_item_status(item)
}
}
\author{
J. Gossa, N. Frerebeau
}
\keyword{internal}
